04-segmentation/requirements.txt
matplotlib~=3.10.0
# install manually because it's CUDA version-specific
# torch~=2.6.0
# torchvision~=0.21
# torchaudio~=2.6.0
torchinfo~=1.8.0
torchmetrics~=1.6.1

imageio~=2.37.0 # for visualizing images in GIF
albumentations~=2.0.4 # for experimenting with image augmentations